+++ contracts/registries/Registry.sol
@@ -74,6 +74,7 @@ abstract contract Registry is IRegistry, ERC721, ENSReverseRegistration, Ownable
 	function setName(address _ens, string calldata _name)
 	external onlyOwner()
 	{
+		// TODO update this as ENSReverseRegistration.ADDR_REVERSE_NODE is not defined on Arbitrum.
 		_setName(IENS(_ens), _name);
 	}